dw: change and fix
change tmp.i to tmp/i
fix add and remove queue mode

+++ ihook/node_matching.sh
@@ -56,6 +56,7 @@ MATCHING()
  if [[ ! -d $LOCK_SPACE/inode ]] ; then
      mkdir -p $LOCK_SPACE/inode >/dev/null 2>&1
      rsync -r --delete --exclude "*ibuild*" --exclude ".svn" $TASK_SPACE/inode.svn/ $LOCK_SPACE/inode/
+     rm -f $LOCK_SPACE/inode/*ibuild*
  fi
 
  for NODE in `cat $IBUILD_ROOT/conf/priority/[$LEVEL_NUMBER]-floor.conf`
@@ -81,6 +82,7 @@ MATCHING()
      chmod 777 -R $TASK_SPACE/inode.svn >/dev/null 2>&1
      mkdir -p $LOCK_SPACE/inode >/dev/null 2>&1
      rsync -r --delete --exclude "*ibuild*" --exclude ".svn" $TASK_SPACE/inode.svn/ $LOCK_SPACE/inode/
+     rm -f $LOCK_SPACE/inode/*ibuild*
  fi
 
  for NODE in `cat $IBUILD_ROOT/conf/priority/[$LEVEL_NUMBER]-floor.conf`
@@ -101,15 +103,20 @@ MATCHING()
 ASSIGN_JOB()
 {
  if [[ `cat $LOCK_SPACE/itask-r$ITASK_REV.jobs | grep $ITASK_REV_MD5 | grep $NODE_MD5` ]] ; then
-     echo "$ITASK_REV|$NODE|$NODE_IP|$ITASK_SPEC_NAME" >>$ITASK_PATH/jobs.txt
-     svn ci -q $IBUILD_SVN_OPTION -m "auto: assign itask-r$ITASK_REV to $NODE" $ITASK_PATH/jobs.txt
-     export QUEUE_SPACE_C=$(svn st $QUEUE_SPACE | grep '^!' | awk -F' ' {'print $3'} | head -n1)
-     svn rm --force $QUEUE_SPACE/$PRIORITY_ITASK_REV
-     if [[ ! -z $QUEUE_SPACE_C ]] ; then
-         svn revert -q $QUEUE_SPACE_C
-         svn rm --force $QUEUE_SPACE_C
-         sleep 1
-     fi
+    echo "$ITASK_REV|$NODE|$NODE_IP|$ITASK_SPEC_NAME" >>$ITASK_PATH/jobs.txt
+    svn ci -q $IBUILD_SVN_OPTION -m "auto: assign itask-r$ITASK_REV to $NODE" $ITASK_PATH/jobs.txt
+    svn rm -q $IBUILD_SVN_OPTION -m "auto: remove $PRIORITY_ITASK_REV" svn://$IBUILD_SVN_SRV/istatus/queue/itask/$PRIORITY_ITASK_REV
+    svn up -q $IBUILD_SVN_OPTION $QUEUE_SPACE
+    if [[ -f $QUEUE_SPACE/$PRIORITY_ITASK_REV ]] ; then
+        svn rm --force $QUEUE_SPACE/$PRIORITY_ITASK_REV
+        rm -f $QUEUE_SPACE/$PRIORITY_ITASK_REV
+    fi
+#     export QUEUE_SPACE_C=$(svn st $QUEUE_SPACE | grep '^!' | awk -F' ' {'print $3'} | head -n1)
+#     if [[ ! -z $QUEUE_SPACE_C ]] ; then
+#         svn revert -q $QUEUE_SPACE_C
+#         svn rm --force $QUEUE_SPACE_C
+#         sleep 1
+#     fi
  fi
  rm -f $LOCK_SPACE/inode/$NODE
  EXIT
